+++ src/azure-cli/azure/cli/command_modules/vm/commands.py
@@ -31,6 +31,10 @@ from azure.cli.command_modules.vm._image_builder import (
 from azure.cli.core.commands import DeploymentOutputLongRunningOperation, CliCommandType
 from azure.cli.core.commands.arm import deployment_validate_table_format, handle_template_based_exception
 
+from azure.cli.command_modules.monitor._exception_handler import monitor_exception_handler
+from azure.cli.command_modules.monitor._client_factory import cf_metric_def
+from azure.cli.core.profiles import ResourceType
+
 
 # pylint: disable=line-too-long, too-many-statements, too-many-locals
 def load_command_table(self, _):
@@ -177,6 +181,16 @@ def load_command_table(self, _):
         operations_tmpl='azure.mgmt.compute.operations#DiskEncryptionSetsOperations.{}',
         client_factory=cf_disk_encryption_set
     )
+    monitor_custom = CliCommandType(
+        operations_tmpl='azure.cli.command_modules.monitor.custom#{}',
+        exception_handler=monitor_exception_handler)
+
+    metric_definitions_sdk = CliCommandType(
+        operations_tmpl='azure.mgmt.monitor.operations#MetricDefinitionsOperations.{}',
+        resource_type=ResourceType.MGMT_MONITOR,
+        client_factory=cf_metric_def,
+        operation_group='metric_definitions',
+        exception_handler=monitor_exception_handler)
 
     with self.command_group('disk', compute_disk_sdk, operation_group='disks', min_api='2017-03-30') as g:
         g.custom_command('create', 'create_managed_disk', supports_no_wait=True, table_transformer=transform_disk_show_table_output, validator=process_disk_or_snapshot_create_namespace)
@@ -457,3 +471,12 @@ def load_command_table(self, _):
 
     with self.command_group('vm monitor log', log_analytics_data_plane_sdk, client_factory=cf_log_analytics_data_plane) as g:
         g.custom_command('show', 'execute_query_for_vm', transform=transform_log_analytics_query_output)
+
+    with self.command_group('vm monitor metrics', custom_command_type=monitor_custom, command_type=metric_definitions_sdk, resource_type=ResourceType.MGMT_MONITOR, operation_group='metric_definitions', min_api='2018-01-01', is_preview=True) as g:
+        from azure.cli.command_modules.monitor.transformers import metrics_table, metrics_definitions_table
+        from azure.cli.core.profiles._shared import APIVersionException
+        try:
+            g.custom_command('tail', 'list_metrics', command_type=monitor_custom, table_transformer=metrics_table)
+            g.command('list-definitions', 'list', table_transformer=metrics_definitions_table)
+        except APIVersionException:
+            pass