+++ src/azure-cli/HISTORY.rst
@@ -23,6 +23,7 @@ Release History
 **Compute**
 
 * Fix `vm create` failure in Azure Stack profile.
+* vm monitor metrics tail/list-definitions: support query metric and list definitions for a vm.
 
 2.0.78
 ++++++